Raise SystemError with message instead of printing
Printing a message to stdout when there is no `rdrand` facility is not a
good idea, as the user of the package may not want unsolicited output.

Instead, they expect a message in the raised exception indicating a
cause. This way they can opt to ignore the error and have their code
deal with the absence of `rdrand` in a graceful way.